+++ packages/checkbox/CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Minor Changes
 
-- [`1e6db1e0`](https://github.com/chakra-ui/chakra-ui/commit/1e6db1e068c6bc0a4c6c6893d1716d284dcbb8f8)
-  [#2876](https://github.com/chakra-ui/chakra-ui/pull/2876) Thanks
+- [`1e6db1e0`](https://github.com/chakra-xui/chakra-xui/commit/1e6db1e068c6bc0a4c6c6893d1716d284dcbb8f8)
+  [#2876](https://github.com/chakra-xui/chakra-xui/pull/2876) Thanks
   [@with-heart](https://github.com/with-heart)! - Deprecated the
   `defaultIsChecked` prop on `Radio` and `Checkbox` in favor of
   `defaultChecked`, which mirrors the default React prop name for this